Add conditional execution to sensitivities

+++ analysis/main.py
@@ -53,19 +53,21 @@ def main():
         print('Success: get_regional')
 
     # Water use sensitivities
-    df_gen_info_water = pd.read_csv(paths['outputs']['gen_info_water'])
-    df_eia_heat_rates = pd.read_excel(
-        paths['inputs']['eia_heat_rates'],
-        skiprows=[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 11],
-        na_values=['Not Available', 'Not Applicable']
-    )
-    df_oc, df_rc = wc.water_use_sensitivies(
-        df_gen_info_water,
-        df_eia_heat_rates
-    )
-    g_oc, g_rc = wc.sensitivity(df_oc, df_rc)
-    g_oc.savefig(paths['outputs']['oc_sensitivity'])
-    g_rc.savefig(paths['outputs']['rc_sensitivity'])
+    if not os.path.exists(g_rc.savefig(paths['outputs']['rc_sensitivity'])):
+        df_gen_info_water = pd.read_csv(paths['outputs']['gen_info_water'])
+        df_eia_heat_rates = pd.read_excel(
+            paths['inputs']['eia_heat_rates'],
+            skiprows=[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 11],
+            na_values=['Not Available', 'Not Applicable']
+        )
+        df_oc, df_rc = wc.water_use_sensitivies(
+            df_gen_info_water,
+            df_eia_heat_rates
+        )
+        g_oc, g_rc = wc.sensitivity(df_oc, df_rc)
+        g_oc.savefig(paths['outputs']['oc_sensitivity'])
+        g_rc.savefig(paths['outputs']['rc_sensitivity'])
+        print('Success: water_use_sensitivies')
 
 
 if __name__ == '__main__':
